feat: add `dry-run` option (#133)
* feat: add `-n --dry-run` option

This will display all projects but will not clean any of them. Useful
for previewing what will be cleaned, in case you are inclined to run
`--all` for example.

* feat: print number of project and bytes found

Now prints `Projects cleaned: 0/7, Bytes deleted: 0.0B / 1.6GiB`, where
`7` is the number of projects found and `1.6GiB` is the combined size of
all projects.

This is particularly useful when running `--dry-run` to see how much
space would be saved.